Use vector of predicates instead of strings as a path description
Update all public functions to have two extra arguments:

  * pred-xs for a collection of functions (f Any x -> Boolean y)
  * detach-children? - boolean

Collection of predicates will be used to check if key/value pair from object
need to be kept in resulted map or not. Predicates are position sencitive -
representing the level of nesting (left to right).

detach-children? option should be used to return lazy collection without parent
elements. E.g. for that json: {"foo": [1, 2, 3]} and for that pred-xs:
[#{"foo"}] with `true` as a detach-children? lazy-seq with '(1 2 3) will return.