Add dependency exclusion for jackson-databind.
The jackson-databind library is not actually used by Cheshire; as far
as I can tell it's not even possible to use from Clojure.

This library is plagued by a nonstop stream of CVEs:

https://www.cvedetails.com/vulnerability-list/vendor_id-15866/product_id-42991/Fasterxml-Jackson-databind.html

After the log4j incident, it seems prudent to reconsider the inclusion
of dependencies that have a lot of "magic" features that aren't
actually used.