Allow parser customizations via protocol
This patch introduces the `IParser` protocol for customizing how values are parsed. It subsumes the
already existing customization options `key-fn`, `array-coerce-fn` and `*use-bigdecimals?*` (see
`backwards-compatible-parser`). For maximal performance, users should reify a parser once and share
it across parser invocations. The default parser is shared by default.

Benchmarks show negligible overhead. To that end, care was taken that the existing inlineable
parsers continue to be inlined when possible.