docs: restore the cheshire puss! :)
I went for a bit taller image than proposed to allow text to right of
image take longer to wrap when screen width is narrower.

Got rid of `<small>` and `padding` html because these are ignored by
github markdown.

I also italicized the quoted text to distinguish it more for other text.
This is especially helpful when the text wraps on narrower screens.